move admin page files around
