+++ emphasis_education_app/src/components/AdminPage/AdminPage.tsx
@@ -2,7 +2,7 @@ import * as React from 'react';
 import { createBottomTabNavigator } from '@react-navigation/bottom-tabs';
 
 import Search from './LiftedSearch';
-import classSearch from './LiftedClassSearch';
+import ClassSearch from './LiftedClassSearch';
 import InviteUser from './InviteUser';
 import AdminChatPicker from './AdminChatPicker';
 
@@ -11,7 +11,6 @@ interface AdminPageProps {
   route: any;
 }
 
-// this should be where admins manage the classes and the students
 const Tab = createBottomTabNavigator();
 
 const AdminPage: React.FC<AdminPageProps> = () => {
@@ -19,7 +18,7 @@ const AdminPage: React.FC<AdminPageProps> = () => {
   return (
     <Tab.Navigator>
       <Tab.Screen name="Users" component={Search} />
-      <Tab.Screen name="Classes" component={classSearch} />
+      <Tab.Screen name="Classes" component={ClassSearch} />
       <Tab.Screen name="Invite User" component={InviteUser} />
       <Tab.Screen name="Admin Chats" component={AdminChatPicker} />
     </Tab.Navigator>
