--- emphasis_education_app/src/components/AdminPage/IndividualResult.tsx
@@ -1,18 +1,0 @@
-import * as React from 'react';
-import { View } from 'react-native';
-import styled from 'styled-components';
-
-export const IndividualResultContainer = styled(View)`
-  padding: 10px 0;
-  display: flex;
-  flexDirection: row;
-  justifyContent: space-between;
-`;
-
-const IndividualResult: React.FC = ({ children }) => (
-  <IndividualResultContainer>
-    {children}
-  </IndividualResultContainer>
-)
-
-export default IndividualResult;
