+++ emphasis_education_app/src/components/AdminPage/Presentational/SearchResults.tsx
@@ -1,27 +1,23 @@
 import * as React from 'react';
 import { Icon } from 'react-native-elements';
 
-import IndividualResult from '../AdminPage/IndividualResult';
+import IndividualResult from '../common';
 import {
   ThemedText,
   GeneralSpacing,
   FONT_STYLES
-} from '../shared';
-import { UserInfoType } from '../../../types/schema-types';
+} from '../../shared';
+import { UserInfoType } from '../../../../types/schema-types';
+
+import { SearchResultsContain } from '../../Search/common';
 
 interface SearchResultsProps {
   searchResults: UserInfoType[];
   navigation: any;
 }
 
-export const SearchResultsContainer: React.FC = ({ children }) => (
-  <GeneralSpacing u={0} r={20} d={0} l={20}>
-    {children}
-  </GeneralSpacing>
-);
-
 const SearchResults: React.FC<SearchResultsProps> = ({ searchResults, navigation }) => (
-  <SearchResultsContainer>
+  <SearchResultsContain>
     {searchResults.map((results, index) => (
       <IndividualResult
         key={index}
@@ -43,7 +39,7 @@ const SearchResults: React.FC<SearchResultsProps> = ({ searchResults, navigation
       </IndividualResult>
       )
     )}
-  </SearchResultsContainer>
+  </SearchResultsContain>
 )
 
 export default SearchResults;
