+++ emphasis_education_app/src/components/AdminPage/common.tsx
@@ -1,8 +1,18 @@
-import {
-  View
-} from 'react-native';
+import * as React from 'react';
+import { View } from 'react-native';
 import styled from 'styled-components';
 
-export const ContentContain = styled(View)`
-  padding-top: 20px;
-`
+export const IndividualResultContainer = styled(View)`
+  padding: 10px 0;
+  display: flex;
+  flexDirection: row;
+  justifyContent: space-between;
+`;
+
+const IndividualResult: React.FC = ({ children }) => (
+  <IndividualResultContainer>
+    {children}
+  </IndividualResultContainer>
+)
+
+export default IndividualResult;
